+++ src/by/it/kharevich/jd02_03/entity/Store.java
@@ -1,0 +1,20 @@
+package by.it.kharevich.jd02_03.entity;
+
+public class Store{
+
+    private final Manager manager;
+    private final Queue queue;
+
+    public Store(Manager manager, Queue queue) {
+        this.manager = manager;
+        this.queue = queue;
+    }
+
+    public Queue getQueue() {
+        return queue;
+    }
+
+    public Manager getManager() {
+        return manager;
+    }
+}
